refactor architecture tests and add more
add arch test

add integration test and refactor TutorialGroup code

fix ArchCondition name

refactor architecture tests and add more

move StudentExamRepository getIndividualDueDate to ExerciseDateServiceual
