+++ src/test/java/de/tum/in/www1/artemis/architecture/ArchitectureTest.java
@@ -1,7 +1,6 @@
 package de.tum.in.www1.artemis.architecture;
 
 import static com.tngtech.archunit.base.DescribedPredicate.and;
-import static com.tngtech.archunit.base.DescribedPredicate.equalTo;
 import static com.tngtech.archunit.base.DescribedPredicate.not;
 import static com.tngtech.archunit.base.DescribedPredicate.or;
 import static com.tngtech.archunit.core.domain.JavaCall.Predicates.target;
@@ -13,9 +12,6 @@ import static com.tngtech.archunit.core.domain.JavaClass.Predicates.simpleName;
 import static com.tngtech.archunit.core.domain.JavaClass.Predicates.simpleNameContaining;
 import static com.tngtech.archunit.core.domain.JavaClass.Predicates.type;
 import static com.tngtech.archunit.core.domain.JavaCodeUnit.Predicates.constructor;
-import static com.tngtech.archunit.core.domain.JavaModifier.ABSTRACT;
-import static com.tngtech.archunit.core.domain.JavaModifier.FINAL;
-import static com.tngtech.archunit.core.domain.properties.CanBeAnnotated.Predicates.annotatedWith;
 import static com.tngtech.archunit.core.domain.properties.HasName.Predicates.nameMatching;
 import static com.tngtech.archunit.core.domain.properties.HasOwner.Predicates.With.owner;
 import static com.tngtech.archunit.core.domain.properties.HasType.Predicates.rawType;
@@ -29,18 +25,13 @@ import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.members;
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.methods;
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses;
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noCodeUnits;
+import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noFields;
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noMethods;
 
-import java.lang.annotation.Annotation;
 import java.nio.file.Files;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import org.apache.commons.lang3.ClassUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.assertj.core.api.Assertions;
 import org.eclipse.jgit.api.Git;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
@@ -50,12 +41,9 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.slf4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.EntityGraph;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.messaging.simp.SimpMessageSendingOperations;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Controller;
@@ -66,13 +54,11 @@ import org.springframework.web.bind.annotation.RestController;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.hazelcast.core.HazelcastInstance;
 import com.tngtech.archunit.base.DescribedPredicate;
+import com.tngtech.archunit.core.domain.Dependency;
 import com.tngtech.archunit.core.domain.JavaAnnotation;
 import com.tngtech.archunit.core.domain.JavaClass;
-import com.tngtech.archunit.core.domain.JavaCodeUnit;
+import com.tngtech.archunit.core.domain.JavaClasses;
 import com.tngtech.archunit.core.domain.JavaEnumConstant;
-import com.tngtech.archunit.core.domain.JavaMethod;
-import com.tngtech.archunit.core.domain.JavaMethodCall;
-import com.tngtech.archunit.core.domain.JavaParameter;
 import com.tngtech.archunit.core.domain.properties.HasAnnotations;
 import com.tngtech.archunit.lang.ArchCondition;
 import com.tngtech.archunit.lang.ArchRule;
@@ -80,14 +66,21 @@ import com.tngtech.archunit.lang.ConditionEvents;
 import com.tngtech.archunit.lang.SimpleConditionEvent;
 import com.tngtech.archunit.library.GeneralCodingRules;
 
-import de.tum.in.www1.artemis.AbstractArchitectureTest;
 import de.tum.in.www1.artemis.config.ApplicationConfiguration;
-import de.tum.in.www1.artemis.security.annotations.enforceRoleInCourse.EnforceRoleInCourse;
-import de.tum.in.www1.artemis.security.annotations.enforceRoleInExercise.EnforceRoleInExercise;
 import de.tum.in.www1.artemis.service.WebsocketMessagingService;
 import de.tum.in.www1.artemis.service.connectors.GitService;
 import de.tum.in.www1.artemis.web.rest.repository.RepositoryResource;
 
+/**
+ * This class contains architecture tests that apply for the whole project.
+ * <p>
+ * For more specific tests please refer to the other architecture test classes:
+ * <ul>
+ * <li>{@link RepositoryArchitectureTest}</li>
+ * <li>{@link ServiceArchitectureTest}</li>
+ * <li>{@link ResourceArchitectureTest}</li>
+ * </ul>
+ */
 class ArchitectureTest extends AbstractArchitectureTest {
 
     @Test
@@ -106,47 +99,6 @@ class ArchitectureTest extends AbstractArchitectureTest {
     }
 
     @Test
-    void testNoWrongServiceImports() {
-        ArchRule rule = noClasses().should().dependOnClassesThat().resideInAnyPackage("org.jvnet.hk2.annotations")
-                .because("this is the wrong service class, use org.springframework.stereotype.Service.");
-        rule.check(allClasses);
-    }
-
-    @Test
-    void testCorrectServiceAnnotation() {
-
-        classes().that().resideInAPackage("de.tum.in.www1.artemis.service..").and().haveSimpleNameEndingWith("Service").and().areNotInterfaces().and().doNotHaveModifier(ABSTRACT)
-                .should().beAnnotatedWith(org.springframework.stereotype.Service.class)
-                .because("services should be consistently managed by Spring's dependency injection container.").check(allClasses);
-
-        classes().that().haveSimpleNameEndingWith("Service").should().notBeAnnotatedWith(Component.class).check(allClasses);
-        classes().that().haveSimpleNameEndingWith("Service").should().notBeAnnotatedWith(RestController.class).check(allClasses);
-        classes().that().haveSimpleNameEndingWith("Service").should().notHaveModifier(FINAL).check(allClasses);
-
-        classes().that().areAnnotatedWith(Service.class).should().haveSimpleNameEndingWith("Service").check(allClasses);
-        classes().that().areAnnotatedWith(Service.class).should().notBeAnnotatedWith(Component.class).check(allClasses);
-        classes().that().areAnnotatedWith(Service.class).should().notBeAnnotatedWith(RestController.class).check(allClasses);
-        classes().that().areAnnotatedWith(Service.class).should().notHaveModifier(FINAL).check(allClasses);
-
-    }
-
-    @Test
-    void testNoUnusedRepositoryMethods() {
-        ArchRule unusedMethods = noMethods().that().areAnnotatedWith(Query.class).and().areDeclaredInClassesThat().areInterfaces().and().areDeclaredInClassesThat()
-                .areAnnotatedWith(Repository.class).should(new ArchCondition<>("not be referenced") {
-
-                    @Override
-                    public void check(JavaMethod javaMethod, ConditionEvents conditionEvents) {
-                        Set<JavaMethodCall> calls = javaMethod.getCallsOfSelf();
-                        if (calls.isEmpty()) {
-                            conditionEvents.add(SimpleConditionEvent.violated(javaMethod, "Method is not used"));
-                        }
-                    }
-                }).because("unused methods should be removed from repositories to keep a clean code base.");
-        unusedMethods.check(productionClasses);
-    }
-
-    @Test
     void testCorrectStringUtils() {
         ArchRule stringUtils = noClasses().should()
                 .dependOnClassesThat(have(simpleName("StringUtils")).and(not(resideInAnyPackage("org.apache.commons.lang3", "org.springframework.util"))));
@@ -161,22 +113,6 @@ class ArchitectureTest extends AbstractArchitectureTest {
         ArchRule noJunitJupiterAssertions = noClasses().should().dependOnClassesThat().haveNameMatching("org.junit.jupiter.api.Assertions");
 
         noJunitJupiterAssertions.check(testClasses);
-    }
-
-    @Test
-    void testNoEntityGraphsOnQueries() {
-        ArchRule noEntityGraphsOnQueries = noMethods().that().areAnnotatedWith(Query.class).and().areDeclaredInClassesThat().areInterfaces().and().areDeclaredInClassesThat()
-                .areAnnotatedWith(Repository.class).should().beAnnotatedWith(EntityGraph.class)
-                .because("Spring Boot 3 ignores EntityGraphs on JPQL queries. You need to integrate a JOIN FETCH into the query.");
-        noEntityGraphsOnQueries.check(productionClasses);
-    }
-
-    @Test
-    void testOnlySpringTransactionalAnnotation() {
-        ArchRule onlySpringTransactionalAnnotation = noMethods().should().beAnnotatedWith(javax.transaction.Transactional.class).orShould()
-                .beAnnotatedWith(jakarta.transaction.Transactional.class)
-                .because("Only Spring's Transactional annotation should be used as the usage of the other two is not reliable.");
-        onlySpringTransactionalAnnotation.check(allClasses);
     }
 
     @Test
@@ -256,14 +192,6 @@ class ArchitectureTest extends AbstractArchitectureTest {
         jsonParser.check(allClasses);
     }
 
-    @Test
-    void testRepositoryParamAnnotation() {
-        var useParamInQueries = methods().that().areAnnotatedWith(Query.class).should(haveAllParametersAnnotatedWithUnless(rawType(Param.class), type(Pageable.class)));
-        var notUseParamOutsideQueries = methods().that().areNotAnnotatedWith(Query.class).should(notHaveAnyParameterAnnotatedWith(rawType(Param.class)));
-        useParamInQueries.check(productionClasses);
-        notUseParamOutsideQueries.check(productionClasses);
-    }
-
     /**
      * Checks that no class directly calls Git.commit(), but instead uses GitService.commit()
      * This is necessary to ensure that committing is identical for all setups, with and without commit signing
@@ -295,118 +223,6 @@ class ArchitectureTest extends AbstractArchitectureTest {
     }
 
     @Test
-    void testJPQLStyle() {
-        var queryRule = methods().that().areAnnotatedWith(Query.class).should(useUpperCaseSQLStyle()).because("@Query content should follow the style guide");
-        queryRule.check(allClasses);
-    }
-
-    @Test
-    void testEnforceRoleInCourseEndpointHasCourseIdParameter() {
-        ArchCondition<JavaMethod> haveParameterWithAnnotation = new ArchCondition<>("have a parameter with EnforceRoleInCourse annotation") {
-
-            @Override
-            public void check(JavaMethod method, ConditionEvents events) {
-                // Get annotation
-                var enforceRoleInCourseAnnotation = getAnnotation(EnforceRoleInCourse.class, method);
-                var courseIdFieldName = enforceRoleInCourseAnnotation.resourceIdFieldName();
-                if (!hasParameterWithName(method, courseIdFieldName)) {
-                    events.add(violated(method, String.format("Method %s does not have a parameter named %s", method.getFullName(), courseIdFieldName)));
-                }
-            }
-        };
-
-        var enforceRoleInCourse = methods().that().areAnnotatedWith(EnforceRoleInCourse.class).or().areMetaAnnotatedWith(EnforceRoleInCourse.class).or().areDeclaredInClassesThat()
-                .areAnnotatedWith(EnforceRoleInCourse.class).and().areDeclaredInClassesThat().areNotAnnotations().should(haveParameterWithAnnotation);
-
-        enforceRoleInCourse.check(productionClasses);
-    }
-
-    @Test
-    void testEnforceRoleInExerciseEndpointHasExerciseIdParameter() {
-        ArchCondition<JavaMethod> haveParameterWithAnnotation = new ArchCondition<>("have a parameter with EnforceRoleInExercise annotation") {
-
-            @Override
-            public void check(JavaMethod method, ConditionEvents events) {
-                // Get annotation
-                var enforceRoleInExerciseAnnotation = getAnnotation(EnforceRoleInExercise.class, method);
-                var exerciseIdFieldName = enforceRoleInExerciseAnnotation.resourceIdFieldName();
-                if (!hasParameterWithName(method, exerciseIdFieldName)) {
-                    events.add(violated(method, String.format("Method %s does not have a parameter named %s", method.getFullName(), exerciseIdFieldName)));
-                }
-            }
-        };
-
-        var enforceRoleInExercise = methods().that().areAnnotatedWith(EnforceRoleInExercise.class).or().areMetaAnnotatedWith(EnforceRoleInExercise.class).or()
-                .areDeclaredInClassesThat().areAnnotatedWith(EnforceRoleInExercise.class).and().areDeclaredInClassesThat().areNotAnnotations().should(haveParameterWithAnnotation);
-
-        enforceRoleInExercise.check(productionClasses);
-    }
-
-    private boolean hasParameterWithName(JavaMethod method, String paramName) {
-        try {
-            var owner = method.getOwner();
-            var javaClass = Class.forName(owner.getFullName());
-            var javaMethod = javaClass.getMethod(method.getName(), method.getRawParameterTypes().stream().map(this::getClassForName).toArray(Class[]::new));
-            return Arrays.stream(javaMethod.getParameters()).anyMatch(parameter -> parameter.getName().equals(paramName));
-        }
-        catch (ClassNotFoundException | NoSuchMethodException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private Class<?> getClassForName(JavaClass paramClass) {
-        try {
-            return ClassUtils.getClass(paramClass.getName());
-        }
-        catch (ClassNotFoundException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private <T extends Annotation> T getAnnotation(Class<T> clazz, JavaMethod javaMethod) {
-        final var method = javaMethod.reflect();
-        T annotation = method.getAnnotation(clazz);
-        if (annotation != null) {
-            return annotation;
-        }
-        for (Annotation a : method.getDeclaredAnnotations()) {
-            annotation = a.annotationType().getAnnotation(clazz);
-            if (annotation != null) {
-                return annotation;
-            }
-        }
-
-        annotation = method.getDeclaringClass().getAnnotation(clazz);
-        if (annotation != null) {
-            return annotation;
-        }
-        for (Annotation a : method.getDeclaringClass().getDeclaredAnnotations()) {
-            annotation = a.annotationType().getAnnotation(clazz);
-            if (annotation != null) {
-                return annotation;
-            }
-        }
-
-        return null;
-    }
-
-    @Test
-    void testTransactional() {
-        var classesPredicated = and(INTERFACES, annotatedWith(Repository.class));
-        var transactionalRule = methods().that().areAnnotatedWith(simpleNameAnnotation("Transactional")).should().beDeclaredInClassesThat(classesPredicated);
-
-        // TODO: In the future we should reduce this number and eventually replace it by transactionalRule.check(allClasses)
-        // The following methods currently violate this rule:
-        // Method <de.tum.in.www1.artemis.service.LectureImportService.importLecture(Lecture, Course)>
-        // Method <de.tum.in.www1.artemis.service.exam.StudentExamService.generateMissingStudentExams(Exam)>
-        // Method <de.tum.in.www1.artemis.service.exam.StudentExamService.generateStudentExams(Exam)>
-        // Method <de.tum.in.www1.artemis.service.programming.ProgrammingExerciseImportBasicService.importProgrammingExerciseBasis(ProgrammingExercise, ProgrammingExercise)>
-        // Method <de.tum.in.www1.artemis.service.tutorialgroups.TutorialGroupsConfigurationService.onTimeZoneUpdate(Course)>
-        var result = transactionalRule.evaluate(allClasses);
-        Assertions.assertThat(result.getFailureReport().getDetails()).hasSize(5);
-    }
-
-    @Test
     void testJsonIncludeNonEmpty() {
         members().that().areAnnotatedWith(JsonInclude.class).should(useJsonIncludeNonEmpty()).check(allClasses);
         classes().that().areAnnotatedWith(JsonInclude.class).should(useJsonIncludeNonEmpty()).check(allClasses);
@@ -431,52 +247,6 @@ class ArchitectureTest extends AbstractArchitectureTest {
         };
     }
 
-    // Custom Predicates for JavaAnnotations since ArchUnit only defines them for classes
-
-    private DescribedPredicate<? super JavaAnnotation<?>> simpleNameAnnotation(String name) {
-        return equalTo(name).as("Annotation with simple name " + name).onResultOf(annotation -> annotation.getRawType().getSimpleName());
-    }
-
-    private DescribedPredicate<? super JavaAnnotation<?>> resideInPackageAnnotation(String packageName) {
-        return equalTo(packageName).as("Annotation in package " + packageName).onResultOf(annotation -> annotation.getRawType().getPackageName());
-    }
-
-    private DescribedPredicate<? super JavaCodeUnit> declaredClassSimpleName(String name) {
-        return equalTo(name).as("Declared in class with simple name " + name).onResultOf(unit -> unit.getOwner().getSimpleName());
-    }
-
-    private ArchCondition<JavaMethod> notHaveAnyParameterAnnotatedWith(DescribedPredicate<? super JavaAnnotation<?>> annotationPredicate) {
-        return new ArchCondition<>("not have parameters annotated with " + annotationPredicate.getDescription()) {
-
-            @Override
-            public void check(JavaMethod item, ConditionEvents events) {
-                boolean satisfied = item.getParameterAnnotations().stream().flatMap(Collection::stream).noneMatch(annotationPredicate);
-                if (!satisfied) {
-                    events.add(violated(item, String.format("Method %s has parameter violating %s", item.getFullName(), annotationPredicate.getDescription())));
-                }
-            }
-        };
-    }
-
-    private ArchCondition<JavaMethod> haveAllParametersAnnotatedWithUnless(DescribedPredicate<? super JavaAnnotation<?>> annotationPredicate,
-            DescribedPredicate<JavaClass> exception) {
-        return new ArchCondition<>("have all parameters annotated with " + annotationPredicate.getDescription()) {
-
-            @Override
-            public void check(JavaMethod item, ConditionEvents events) {
-                boolean satisfied = item.getParameters().stream()
-                        // Ignore annotations of the Pageable parameter
-                        .filter(javaParameter -> !exception.test(javaParameter.getRawType())).map(JavaParameter::getAnnotations)
-                        // Else, one of the annotations should match the given predicate
-                        // This allows parameters with multiple annotations (e.g. @NonNull @Param)
-                        .allMatch(annotations -> annotations.stream().anyMatch(annotationPredicate));
-                if (!satisfied) {
-                    events.add(violated(item, String.format("Method %s has parameter violating %s", item.getFullName(), annotationPredicate.getDescription())));
-                }
-            }
-        };
-    }
-
     private static ArchCondition<JavaClass> beProfileAnnotated() {
         return new ArchCondition<>("be annotated with @Profile") {
 
@@ -491,32 +261,29 @@ class ArchitectureTest extends AbstractArchitectureTest {
         };
     }
 
-    // See https://openjpa.apache.org/builds/1.2.3/apache-openjpa/docs/jpa_langref.html#jpa_langref_from_identifiers
-    private static final Set<String> SQL_KEYWORDS = Set.of("SELECT", "UPDATE", "SET", "DELETE", "DISTINCT", "EXISTS", "FROM", "WHERE", "LEFT", "OUTER", "INNER", "JOIN", "FETCH",
-            "TREAT", "AND", "OR", "AS", "ON", "ORDER", "BY", "ASC", "DSC", "GROUP", "COUNT", "SUM", "AVG", "MAX", "MIN", "IS", "NOT", "FALSE", "TRUE", "NULL", "LIKE", "IN",
-            "BETWEEN", "HAVING", "EMPTY", "MEMBER", "OF", "UPPER", "LOWER", "TRIM");
+    @Test
+    void testNoRestControllersImported() {
+        final var exceptions = new String[] { "AccountResourceIntegrationTest", "AndroidAppSiteAssociationResourceTest", "AppleAppSiteAssociationResourceTest" };
+        final var classes = classesExcept(allClasses, exceptions);
+        classes().should(IMPORT_RESTCONTROLLER).check(classes);
+    }
 
-    private ArchCondition<JavaMethod> useUpperCaseSQLStyle() {
-        return new ArchCondition<>("have keywords in upper case") {
+    private static final ArchCondition<JavaClass> IMPORT_RESTCONTROLLER = new ArchCondition<>("not import RestController") {
 
-            @Override
-            public void check(JavaMethod item, ConditionEvents events) {
-                var queryAnnotation = item.getAnnotations().stream().filter(simpleNameAnnotation("Query")).findAny();
-                if (queryAnnotation.isEmpty()) {
-                    return;
-                }
-                Object valueProperty = queryAnnotation.get().getExplicitlyDeclaredProperty("value");
-                if (!(valueProperty instanceof String query)) {
-                    return;
-                }
-                String[] queryWords = query.split("[\\r\\n ]+");
+        @Override
+        public void check(JavaClass item, ConditionEvents events) {
+            item.getDirectDependenciesFromSelf().stream().map(Dependency::getTargetClass).filter(targetClass -> targetClass.isAnnotatedWith(RestController.class))
+                    .forEach(targetClass -> {
+                        events.add(violated(item, "%s imports the RestController %s".formatted(item.getName(), targetClass.getName())));
+                    });
+        }
+    };
 
-                for (var word : queryWords) {
-                    if (SQL_KEYWORDS.contains(word.toUpperCase()) && !StringUtils.isAllUpperCase(word)) {
-                        events.add(violated(item, "In the Query of %s the keyword \"%s\" should be written in upper case.".formatted(item.getFullName(), word)));
-                    }
-                }
-            }
-        };
+    @Test
+    void shouldNotUserAutowiredAnnotation() {
+        ArchRule rule = noFields().should().beAnnotatedWith(Autowired.class).because("fields should not rely on field injection via @Autowired");
+        final var exceptions = new String[] { "PublicResourcesConfiguration", "QuizProcessCacheTask", "QuizStartTask" };
+        JavaClasses classes = classesExcept(productionClasses, exceptions);
+        rule.check(classes);
     }
 }
